Fix handling of utf-8 encoding in AtteanX::Parser::SPARQLXML.
This issue occurred because the XML parser requried data to be utf-8 encoded
octets, but in some cases a filehandle with a PerlIO utf-8 decoding layer
applied was being used to read the data. The fix is to check if such a layer
is present on the filehandle, and re-encode data as necessary.